Added function that takes text as new Object name
The reason my code wasn't working before was that I was trying to access
a div that didn't yet exist. To fix that, I added the div event listener
inside of the button event listener and it worked.

Need to be sure to take my time looking at the process of why an error
occurs. Load it in steps, line by line.